chore: show a dialog when clicking on now playing details
