+++ data/com.vixalien.muzika.data.gresource.xml.in
@@ -21,18 +21,18 @@
     <file preprocess="xml-stripblanks">ui/components/navbar/index.ui</file>
     <file preprocess="xml-stripblanks">ui/components/navbar/title.ui</file>
     <file preprocess="xml-stripblanks">ui/components/paginator.ui</file>
-    <file preprocess="xml-stripblanks">ui/components/player/now-playing/details.ui</file>
+    <file preprocess="xml-stripblanks">ui/components/player/now-playing/details/lyrics.ui</file>
+    <file preprocess="xml-stripblanks">ui/components/player/now-playing/details/queue.ui</file>
+    <file preprocess="xml-stripblanks">ui/components/player/now-playing/details/queueitem.ui</file>
+    <file preprocess="xml-stripblanks">ui/components/player/now-playing/details/related.ui</file>
+    <file preprocess="xml-stripblanks">ui/components/player/now-playing/cover.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/now-playing/view.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/video/controls.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/video/view.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/video/volume-controls.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/full.ui</file>
-    <file preprocess="xml-stripblanks">ui/components/player/lyrics.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/mini.ui</file>
     <file preprocess="xml-stripblanks">ui/components/player/preview.ui</file>
-    <file preprocess="xml-stripblanks">ui/components/player/queue.ui</file>
-    <file preprocess="xml-stripblanks">ui/components/player/queueitem.ui</file>
-    <file preprocess="xml-stripblanks">ui/components/player/related.ui</file>
     <file preprocess="xml-stripblanks">ui/components/playlist/add-to-playlist-item.ui</file>
     <file preprocess="xml-stripblanks">ui/components/playlist/bar.ui</file>
     <file preprocess="xml-stripblanks">ui/components/playlist/edit.ui</file>
