--- data/ui/components/player/now-playing/details.blp
@@ -1,48 +1,0 @@
-using Gtk 4.0;
-using Gdk 4.0;
-using Adw 1;
-
-template $PlayerNowPlayingDetails : Adw.NavigationPage {
-  title: _("Queue");
-
-  Adw.ToolbarView toolbar_view {
-    bottom-bar-style: raised;
-    reveal-bottom-bars: false;
-
-    [top]
-    Adw.HeaderBar headerbar {
-      title-widget: Adw.ViewSwitcher {
-        stack: stack;
-        policy: wide;
-      };
-    }
-
-    Adw.ViewStack stack {
-      vhomogeneous: false;
-
-      Adw.ViewStackPage {
-        name: "queue";
-        icon-name: "playlist2-symbolic";
-        title: _("Queue");
-        child: $QueueView {};
-      }
-
-      Adw.ViewStackPage lyrics_page {
-        name: "lyrics";
-        icon-name: "subtitles-symbolic";
-        title: _("Lyrics");
-        child: $LyricsView {};
-      }
-
-      Adw.ViewStackPage related_page {
-        name: "related";
-        icon-name: "navigate-symbolic";
-        title: _("Related");
-        child: $RelatedView {};
-      }
-    }
-
-    [bottom]
-    $MiniPlayerView {}
-  }
-}
