+++ data/ui/components/player/now-playing/details/lyrics.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $LyricsView : Stack {
+template $MuzikaNPLyrics : Stack {
   Adw.StatusPage no_lyrics {
     icon-name: "heart-broken-symbolic";
     title: _("Lyrics are not available");
