+++ data/ui/components/player/now-playing/details/queue.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $QueueView : Stack {
+template $MuzikaNPQueue : Stack {
   Adw.StatusPage no_queue {
     icon-name: "playlist2-symbolic";
     title: _("The Queue is empty");
