+++ data/ui/components/player/now-playing/details/queueitem.blp
@@ -1,6 +1,6 @@
 using Gtk 4.0;
 
-template $QueueItem : Box {
+template $MuzikaNPQueueItem : Box {
   margin-top: 6;
   margin-bottom: 6;
   margin-start: 9;
