+++ data/ui/components/player/now-playing/details/related.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $RelatedView : Stack {
+template $MuzikaNPRelated : Stack {
   Adw.StatusPage no_related {
     icon-name: "heart-broken-symbolic";
     title: _("No related music");
