+++ data/ui/meson.build
@@ -15,18 +15,18 @@ blueprints = custom_target('blueprints',
     'components/navbar/index.blp',
     'components/navbar/title.blp',
     'components/paginator.blp',
-    'components/player/now-playing/details.blp',
+    'components/player/now-playing/details/lyrics.blp',
+    'components/player/now-playing/details/queue.blp',
+    'components/player/now-playing/details/queueitem.blp',
+    'components/player/now-playing/details/related.blp',
+    'components/player/now-playing/cover.blp',
     'components/player/now-playing/view.blp',
     'components/player/video/controls.blp',
     'components/player/video/view.blp',
     'components/player/video/volume-controls.blp',
     'components/player/full.blp',
-    'components/player/lyrics.blp',
     'components/player/mini.blp',
     'components/player/preview.blp',
-    'components/player/queue.blp',
-    'components/player/queueitem.blp',
-    'components/player/related.blp',
     'components/playlist/add-to-playlist-item.blp',
     'components/playlist/bar.blp',
     'components/playlist/edit.blp',
