+++ data/ui/window.blp
@@ -5,33 +5,14 @@ template $MuzikaWindow : Adw.ApplicationWindow {
   default-width: 1000;
   default-height: 800;
   width-request: 360;
-  height-request: 294;
-
-  Adw.Breakpoint {
-    condition ("max-width: 1000sp")
-
-    setters {
-      now_playing_split_view.sidebar-width-fraction: 0.6;
-    }
-  }
-
-  Adw.Breakpoint {
-    condition ("max-width: 800sp")
-
-    setters {
-      now_playing_split_view.sidebar-width-fraction: 0.5;
-    }
-  }
+  height-request: 400;
 
   Adw.Breakpoint {
     condition ("max-width: 600sp")
 
     setters {
       split_view.collapsed: true;
-      now_playing_split_view.collapsed: true;
-      now_playing_details.header-title: null;
-      now_playing_details.show-mini-player: true;
-      now_playing_view.switcher-visible: true;
+      now_playing_view.show-mini: true;
     }
     
     apply => $on_breakpoint_apply();
@@ -67,17 +48,7 @@ template $MuzikaWindow : Adw.ApplicationWindow {
 
         StackPage {
           name: "now-playing";
-          child: Adw.NavigationSplitView now_playing_split_view {
-            sidebar-width-fraction: 0.7;
-            min-sidebar-width: 0;
-            max-sidebar-width: 9999;
-
-            sidebar: $PlayerNowPlayingView now_playing_view {
-              switcher-stack: bind now_playing_details.stack;
-            };
-
-            content: $PlayerNowPlayingDetails now_playing_details {};
-          };
+          child: $MuzikaNPView now_playing_view {};
         }
 
         StackPage {
