+++ src/components/player/now-playing/details/lyrics.ts
@@ -36,12 +36,12 @@ export class TimedLyricLineRow extends Gtk.ListBoxRow {
   }
 }
 
-export class LyricsView extends Gtk.Stack {
+export class MuzikaNPLyrics extends Gtk.Stack {
   static {
     GObject.registerClass({
-      GTypeName: "LyricsView",
+      GTypeName: "MuzikaNPLyrics",
       Template:
-        "resource:///com/vixalien/muzika/ui/components/player/lyrics.ui",
+        "resource:///com/vixalien/muzika/ui/components/player/now-playing/details/lyrics.ui",
       InternalChildren: [
         "no_lyrics",
         "loading",
@@ -177,7 +177,6 @@ export class LyricsView extends Gtk.Stack {
 
     if (this.pending_animation) {
       this.pending_animation.pause();
-
       this.pending_animation = null;
     }
 
@@ -197,7 +196,7 @@ export class LyricsView extends Gtk.Stack {
         this._timed_window,
         this._timed_window.vadjustment.value,
         Math.min(Math.max(scroll_to, 0), this._timed_window.vadjustment.upper),
-        LyricsView.ANIMATION_DURATION,
+        MuzikaNPLyrics.ANIMATION_DURATION,
         property_target,
       );
